% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agaricales_enrich.R
\docType{data}
\name{agaricales_enrich}
\alias{agaricales_enrich}
\title{Sample  of Agaricales enrichment data set}
\format{
A data frame with 10000 rows and 18 variables:
\describe{
\item{new_full_name}{up-to-date scientific name}
\item{new_name}{up-to-date canonical name}
\item{new_author}{up-to-date authorship}
\item{new_kingdom}{up-to-date kingdom}
\item{new_phylum}{up-to-date phylum}
\item{new_order}{up-to-date order}
\item{new_family}{up-to-date family}
\item{new_genus}{up-to-date genus}
\item{new_specific_epithet}{up-to-date specific epithet}
\item{freq}{total frequency of taxon in the data set}
\item{trait_freq}{freqeuncy of taxon in the fire-associated records data set}
\item{trait_ratio}{fire-association enrichment factor}
\item{max_bias}{max bias for taxon among total records}
\item{coll_groups}{number of collector groups for taxon among total records}
\item{max_bias_t}{max bias for taxon among trait records}
\item{coll_groups_t}{number of collector groups for taxon among trait records}
}
}
\source{
\url{https://mycoportal.org}
}
\usage{
agaricales_enrich
}
\description{
Enrichment data set generated using global Agaricales fungal records downloaded from the MyCoPortal (https://mycoportal.org) on Jan. 20, 2021.
Downloaded as tab-delimited file in ISO-8859-1 encoding.
Taxon names updated on Jan. 25, 2021.
Enrichment calculated on Mar. 4, 2021. Trait analyzed: fire-association.
}
\keyword{datasets}
