% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwca.R
\name{as_dwca}
\alias{as_dwca}
\title{Create a DWCA object}
\usage{
as_dwca(data)
}
\arguments{
\item{data}{A data.frame with required Darwin Core fields. ("scientificName",
"scientificNameAuthorship", "eventDate", "country", "stateProvince",
"decimalLatitude", "decimalLongitude", "habitat", "occurrenceRemarks",
"associatedTaxa")}

\item{drop_cols}{Logical. Drop all nonrequired columns. Default: TRUE.}
}
\value{
An object of class 'dwca' (and 'data.frame').
}
\description{
Validates and assigns the S3 class 'dwca' to a data.frame.
}
\examples{
data <- data.frame(scientificName = "Quercus alba", occurrenceID = "123")
dw <- as_dwca(data, required_cols = c("scientificName", "occurrenceID"))
class(dw)  # c("dwca", "data.frame")
}
