% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_funguild.R
\name{assign_funguild}
\alias{assign_funguild}
\title{Assign FUNGuild data}
\usage{
assign_funguild(
  data,
  url = "http://www.stbates.org/funguild_db.php",
  tax_cols = c("phylum_pres", "class_pres", "order_pres", "family_pres", "genus_pres",
    "species_pres")
)
}
\arguments{
\item{data}{Data.frame.}

\item{url}{Character. URL where FUNGuild database lives. Current default: http://www.stbates.org/funguild_db.php.}

\item{tax_cols}{Character. Column names containing taxonomic variables. Names must be in descending order of taxonomic rank. Default is c("phylum_pres", "class_pres", "order_pres", "family_pres", "genus_pres", "species_pres").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Input data.frame with FUNGuild assignments appended.
}\if{html}{\out{</div>}}
}
\description{
Assigns FUNGuild data (e.g. trophic mode, functional guild, etc) to taxa listed in a
data.frame using the FUNGuild database (PLEASE CITE:Nguyen et al. 2016).
}
\examples{
library(fungarium)
data(fomitopsidaceae_clean_taxonomy) #import sample data set
fomitopsidaceae_w_funguild <- assign_funguild(fomitopsidaceae_clean_taxonomy)

}
\references{
\enumerate{
\item Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J, Schilling JS, Kennedy PG. 2016. FUNGuild: An open annotation tool for parsing fungal community datasets by ecological guild. \emph{Fungal Ecology}, 20: 241-248. doi:10.1016/j.funeco.2015.06.006
}
}
