% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grid.R
\name{assign_grid}
\alias{assign_grid}
\title{Overlay grid with equal area cells onto map}
\usage{
assign_grid(
  data,
  lat_col = "lat_parsed",
  lon_col = "lon_parsed",
  size = 10000L,
  proj = "cea"
)
}
\arguments{
\item{data}{Data.frame.}

\item{lat_col}{character. Name of column that contains latitude values.}

\item{lon_col}{Character. Name of column that contains longitude values.}

\item{size}{Integer. Grid cell area (in square kilometers). Default is 10000.}

\item{proj}{Character. Equal area map projection to use. Options: 'cea'
(cylindrical equal area) or 'eqearth' (equal earth). Default is 'cea'.}
}
\value{
List containing the following elements:
\describe{
\item{\code{data}}{Input data.frame with appended "\code{grid_cell_id}".}
\item{\code{grid_ref}}{Data.frame linking "\code{grid_cell_id}" to corresponding grid cell geometries. "\code{grid_cell_size}" is in sqauare kilometers}
}
}
\description{
Overlays hexagonal grid cells of specified area onto equal area world map.
}
\note{
For more info on the Equal Earth projection see:
https://proj.org/en/stable/operations/projections/eqearth.html.
For cylindrical equal area see: https://proj.org/en/stable/operations/projections/cea.html.
}
\examples{
library(fungarium)
data(fomitopsidaceae_clean_geo) #import sample data set
fomitopsidaceae_w_funguild <- assign_grid(fomitopsidaceae_clean_geo)
}
