% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate abundance}
\usage{
calc_abundance(
  data,
  of = "species_pres",
  per = c("year_parsed", "grid_cell_id"),
  relative = FALSE
)
}
\arguments{
\item{data}{Data.frame.}

\item{of}{Character. Name of column containing the variables for which you are calculating abundance. Default is "species_pres".}

\item{per}{Character. Vector of column names. Abundance for each "\code{of}" variable will be calculated for each combination of "\code{per}" variables. Default is c("year_parsed", "grid_cell_id").}

\item{relative}{Logical. Convert abundance to relative abundance.}
}
\value{
Data.frame containing abundance values for each "\code{of}" variable (as
columns) and each combination of "\code{per}" variables as rows. Total occurrences
for each row are recorded in "\code{total_occ}" column.
}
\description{
Calculate abundances for a particular variable (e.g., species) based on
specified grouping variables (e.g., year, country).
}
\examples{
library(fungarium)
data(fomitopsidaceae_fully_cleaned) #import sample data set
grid <- assign_grid(fomitopsidaceae_fully_cleaned)
grid$data <- grid$data[!is.na(grid$data$grid_cell_id)&!is.na(grid$data$year_parsed),]
abundance <- calc_abundance(grid$data)
}
