% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dates.R
\name{clean_dates}
\alias{clean_dates}
\title{Clean heterogeneous dates in occurrence data sets}
\usage{
clean_dates(dates)
}
\arguments{
\item{dates}{Character. Dates with varying formats (e.g., "1990-01-24",
"13-Jan-2024", "03/12/2003").}
}
\value{
Data.frame with the following output fields.

\describe{
\item{\code{date_raw}}{Character. Input date.}
\item{\code{detected_format}}{Character. Detected date format (e.g., "Y", "Ym", or "Ymd").}
\item{\code{parsing_error}}{Character. Error that prevented successful parsing.}
\item{\code{year_parsed}}{Integer. Year parsed into harmonized format (YYYY).}
\item{\code{month_parsed}}{Integer. Month parsed into harmonized format (MM).}
\item{\code{day_parsed}}{Integer. Day parsed into harmonized format (DD).}
\item{\code{date_parsed}}{Date. Dates parsed into harmonized format (YYYY-MM-DD).}
}
}
\description{
Parses dates with varying formats into a consistent format and extracts year,
month, and day.
}
\note{
Full dates assumed to be "Ymd" or "dmY"
format, not "Ydm", or "mdY". Ex: 01/10/1990 is assumed to be October 10,
1990, not January 10, 1990, and "2020-03-12" is assumed to be March 12, 2020,
not December 3, 2020.
}
\examples{
library(fungarium)
data(agaricales_updated) #import sample data set
clean_dates <- clean_dates(agaricales_updated$eventDate) #clean dates

}
