% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_geo.R
\name{clean_geo}
\alias{clean_geo}
\title{Clean and harmonize heterogeneous geographic data (lat/lon/country)}
\usage{
clean_geo(
  data,
  tests = c("zero", "equal", "country", "centroid"),
  centroid_dis = 100L
)
}
\arguments{
\item{data}{\code{dwca} object.}

\item{tests}{Character. Coordinate cleaning tests to perform. Options include: "zero", "equal", "country", "centroid". Default is all tests.}

\item{centroid_dis}{Integer. Distance threshold (in meters) to use for the centroid test. Default is 100.}

\item{round_digits}{Integer. Number of decimal places to use for rounding coordinates. Default is 4. If NULL, no rounding is performed.}
}
\value{
Data.frame with the following output fields.

\describe{
\item{\code{lat_raw}}{Character. Input latitude.}
\item{\code{lon_raw}}{Character. Input longitude.}
\item{\code{country_raw}}{Character. Input country.}
\item{\code{lat_parsed}}{Numeric. Parsed decimal latitude.}
\item{\code{lon_parsed}}{Numeric. Parsed decimal longitude.}
\item{\code{country_parsed}}{Character. Parsed country. Synonyms are harmonized.}
\item{\code{lat_res}}{Integer. Latitude resolution based on number of decimal places.}
\item{\code{lon_res}}{Integer. Longitude resolution based on number of decimal places.}
\item{\code{coordinate_error}}{Character. First error deteced when cleaning coordinates.}
}
\description{
Parse varying formats/spellings for geographic data into a consistent formats
and check the validity of lat/lon coordinates using various tests
}
\details{
The following tests are automatically done:
\describe{
\item{\code{non_numeric}}{lat or lon are not numeric or cannot be converted to numeric}
\item{\code{out_of_bounds}}{lat or lon are out of bounds (i.e., lat>90, lat<-90, lon>180, lon<-180)}
\item{\code{null}}{lat or lon are null}
}
The following tests can be selected:
\describe{
\item{\code{zero}}{lat and lon are both zero}
\item{\code{equal}}{lat and lon are equal}
}
}
\note{
Input coordinates are assumed to use WGS84 coordinate reference system.
}
\examples{
library(fungarium)
data(agaricales_updated) #import sample data set
agaricales_geo_clean <- clean_geo(agaricales_updated) #clean coordinates

}
