% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_taxonomy.R
\name{clean_taxonomy}
\alias{clean_taxonomy}
\title{Clean and harmonize taxon names}
\usage{
clean_taxonomy(
  data,
  name_col = "scientificName",
  author_col = "scientificNameAuthorship",
  kingdom = "Fungi",
  refresh_db = FALSE,
  db_url = "https://download.catalogueoflife.org/col/annual/2024_dwca.zip",
  threads = 1L
)
}
\arguments{
\item{data}{\code{Data.frame}.}

\item{name_col}{Character. Name of column containing taxon names.}

\item{author_col}{Character. Name of column containing author names.}

\item{kingdom}{Character. \code{Fungi} or \code{Plantae}. Used to speed up taxon matching. Default: \code{Fungi}.}

\item{refresh_db}{Logical. Should the COL database be refreshed. Default: \code{FALSE}. Note, database is always downloaded during the first execution of this function after package installation.}

\item{db_url}{Character. URL for COL database. Default: \verb{https://download.catalogueoflife.org/col/annual/2024_dwca.zip}.}
}
\value{
Input \code{data.frame} with the following output fields appended.

\describe{
\item{\code{kingdom_pres}}{Character. Accepted kingdom at present.}
\item{\code{phylum_pres}}{Character. Accepted phylum at present.}
\item{\code{class_pres}}{Character. Accepted class at present.}
\item{\code{order_pres}}{Character. Accepted order at present.}
\item{\code{family_pres}}{Character. Accepted family at present.}
\item{\code{genus_pres}}{Character. Accepted genus at present.}
\item{\code{specific_epithet_pres}}{Character. Accepted specific epithet at present.}
\item{\code{species_pres}}{Character. Accepted species at present.}
\item{\code{species_authority_pres}}{Character. Accepted authority at present.}
\item{\code{match_type}}{Character. 'EXACT' vs 'FUZZY'. See full description below.}
\item{\code{match_score}}{Integer. Confidence score (0-100) for match quality.}
}
}
\description{
Cleans and updates taxon names of fungi and plants based on \href{https://www.catalogueoflife.org/}{Catalogue of Life} data.
Synonyms are updated to currently accepted names and current taxonomic hierarchies are assigned.
}
\note{
Match type may include: 'EXACT-FULL' when full taxon name (species AND
authority) is matched perfectly with record in COL data, 'EXACT-PARTIAL'
when species name without authority is matched perfectly with record in COL
data, 'FUZZY-FULL' when full taxon name (species AND authority) is fuzzy
matched (based on string alignment algorithms) with
record in COL data, FUZZY-PARTIAL' when species name without authority is
fuzzy matched with record in COL data. 'FUZZY' matching is done when 'EXACT'
matches are not found and come with a match score indicating match quality.
Fuzzy matching is done for both full and partial names and whichever
produces the highest match score is reported.
}
\examples{
library(fungarium)
data(fomitopsidaceae) #import sample data set
clean_tax <- clean_taxonomy(fomitopsidaceae) #clean taxonomy

}
