% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_clean.R
\name{coord_clean}
\alias{coord_clean}
\title{Clean lat/long coordinates in fungal collections data sets}
\usage{
coord_clean(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country",
  tests = c("zero", "equal", "countries", "centroids", "capitals"),
  centroid_dis = 100,
  show_status = T,
  round_digits = 4
)
}
\arguments{
\item{data}{Data.frame of fungal collections data set (e.g. MyCoPortal, GBIF)}

\item{lat}{Character string specifying the decimal latitude column. Default is "decimalLatitude" (i.e., the Darwin Core standard).}

\item{lon}{Character string specifying the decimal longitude column. Default is "decimalLongitude" (i.e., the Darwin Core standard).}

\item{tests}{Character vector specifying the coordinates cleaning tests to perform (i.e., "zero", "equal", "countries", "centroids"). Default is c("zero", "equal", "countries", "centroids").}
}
\value{
Data.frame containing the input data set with the following output fields appended.
Records with dates that could not be parsed are removed.
\item{date_fixed}{NULL}
\item{parsed_date}{NULL}
\item{parsed_format}{NULL}
\item{year_fixed}{NULL}
\item{month_fixed}{NULL}
\item{day_fixed}{NULL}
}
\description{
Parses variably formatted dates into a consistent format of specified resolution (e.g., year, month, day) and removes problematic dates that cannot be parsed.
}
\note{
Full dates other than the "Ymd" format are all assumed to be in the "dmY" format, not "mdY". Ex: 01/10/1990 is assumed to be October 10, 1990, not January 10, 1990.
}
\examples{
library(fungarium)
data(strophariaceae) #import sample data set
data <- date_clean(data) #clean dates

}
