% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_clean.R
\name{date_clean}
\alias{date_clean}
\title{Clean dates in occurrence data sets}
\usage{
date_clean(data, date_col = "eventDate", date_res = "year")
}
\arguments{
\item{data}{Data.frame of occurrence data set (e.g. MyCoPortal, GBIF)}

\item{date_col}{Character string specifying the date column. Default is "eventDate" (i.e., the Darwin Core standard).}

\item{date_res}{Character string specifying the desired resolution for parsed dates (i.e., "year", "month", "day"). Default is "year".}
}
\value{
Data.frame containing the input data set with the following output fields appended.
Records with dates that could not be parsed or do not match the specified resolution are removed.
\describe{
\item{\code{date_fixed}}{Variabe date formats are transformed into standard YYYY-MM-DD format. Dates limited to year or month resolution (e.g., 1990-00-00 or 1990-01-00), are transformed to YYYY (e.g., 1990) or YYYY-MM (e.g., 1990-01) format respectively.}
\item{\code{parsed_format}}{Detected date format (e.g., "Y", "Ym", or "Ymd").}
\item{\code{parsed_date}}{"date_fixed" values are parsed into standard POSIXct date-time objects using \code{lubridate::parse_date_time}}
\item{\code{year_fixed}}{The collection year value after cleaning and parsing dates.}
\item{\code{month_fixed}}{The collection month value after cleaning and parsing dates.}
\item{\code{day_fixed}}{The collection day value after cleaning and parsing dates.}
}
}
\description{
Parses dates with varying format into a consistent format of specified resolution
(e.g., year, month, day) and removes problematic dates that cannot be parsed or
do not match the specified resolution.
}
\note{
Full dates other than the "Ymd" format are all assumed to be in the "dmY"
format, not "mdY". Ex: 01/10/1990 is assumed to be October 10, 1990, not January 10, 1990.
}
\examples{
library(fungarium)
data(agaricales_updated) #import sample data set
clean <- date_clean(agaricales_updated) #clean dates

}
