% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fomitopsidaceae.R
\docType{data}
\name{fomitopsidaceae}
\alias{fomitopsidaceae}
\title{Fomitopsidaceae occurrence records for the United States}
\format{
Data.frame with Darwin Core Archive fields.
}
\source{
\url{https://mycoportal.org/portal/}
}
\usage{
fomitopsidaceae
}
\description{
Data downloaded from MyCoPortal on 6/29/2026 in Darwin Core Archive format.
}
\keyword{datasets}
