% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fips.R
\name{get_fips}
\alias{get_fips}
\title{Assign fips codes}
\usage{
get_fips(
  df,
  state_col = "stateProvince",
  county_col = "county",
  counties = TRUE
)
}
\arguments{
\item{df}{Dataframe containing a column with state names and (optionally) a column of county names.}

\item{state_col}{Character string specifying state column name. Defaults to "stateProvince" (name used by MyCoPortal).}

\item{county_col}{Character string specifying county column name. Defaults to "county" (name used by MyCoPortal).}

\item{counties}{Logical, if TRUE fips codes are retruned for states AND counties in the input dataframe}
}
\value{
Returns input dataframe with fips codes appended:
\item{state_fips}{Two digit state fips code.}
\item{county_fips}{Five digit county fips code.}
\item{state_matchtype}{}
\item{state_conf}{}
\item{county_matchtype}{}
\item{county_conf}{}
}
\description{
Assigns fips codes to states or counties.
}
\details{
Fips codes are assigned based on reference data from...\cr
\cr
When assigning fips codes, approximate name matches are possible
when names have variable nomenclature (e.g. "Anoka", "Anoka Co.", "Anoka County") or are simply
mispelled (e.g "Florda"). See "Return" sections for ways to validate the quality of
approximate matches.
}
\examples{

}
