% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_grid.R
\name{hex_grid}
\alias{hex_grid}
\title{Overlay grid on map}
\usage{
hex_grid(size, map)
}
\arguments{
\item{size}{Integer specifying grid cell area (in square kilometers).}

\item{map}{Map shapefile (sf object). Must use an equal area projection (e.g., "cea")}
}
\value{
sf object containing:
\describe{
\item{\code{x}}{grid cell geometry}
\item{\code{hex}}{hex number. Each grid cell has a unique number}
\item{\code{size}}{the area (in sqauare kilometers) of each grid cell}
}
}
\description{
Overlays hexagonal grid cells of specified area onto map.
}
\examples{
library(fungarium)

#import world shape file
shp <- rnaturalearth::ne_countries('large', returnclass = "sf")

#transform to cylindrical projection (equal area)
shp <- sf::st_transform(shp, crs = "+proj=cea +ellps=WGS84 +datum=WGS84")

#get grid
grid <- hex_grid(80000, shp)
}
