% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycoportal_tab.R
\name{mycoportal_tab}
\alias{mycoportal_tab}
\title{Retrieve MyCoPortal datasets}
\usage{
mycoportal_tab(
  taxon,
  country = "",
  state = "",
  county = "",
  locality = "",
  elevation_from = "",
  elevation_to = "",
  host = "",
  taxon_type = "1",
  north_lat = "",
  south_lat = "",
  west_lon = "",
  east_lon = "",
  point_lat = "",
  point_lon = "",
  radius = "",
  collector = "",
  collector_num = "",
  coll_date1 = "",
  coll_date2 = "",
  verbose = TRUE,
  port = 4445L,
  browserName = "chrome"
)
}
\arguments{
\item{taxon}{Character string specifying the taxon name (e.g., species name, family name or higher taxon). To query higher taxa, e.g. on order level, I recommend using \code{\link{mycoportal_hightax}}}

\item{country}{Character string specifying country, e.g., "USA"}

\item{state}{Character string specifying state, e.g., "Massachusetts"}

\item{county}{Character string specifying county, e.g., "Worcester"}

\item{locality}{Character string specifying locality, e.g., "Harvard Forest"}

\item{elevation_from}{Character string, meter, e.g., "1000"}

\item{elevation_to}{Character string, meter}

\item{host}{Character string specifying host species, e.g., "Betula alba"}

\item{taxon_type}{Integer, one of 1 to 5 representing "Family or Scientific Name", "Scientific Name only", "Family Only", "Higher Taxonomy", Common Name"}

\item{north_lat}{Character string, coordinate e.g., "45"}

\item{south_lat}{Character string, coordinate}

\item{west_lon}{Character string, coordinate, e.g., "-72"}

\item{east_lon}{Character string, coordinate}

\item{point_lat}{Character string, coordinate}

\item{point_lon}{Character string, coordinate}

\item{radius}{Character string, km, e.g., "50"}

\item{collector}{Character string specifying collector name}

\item{collector_num}{Character string specifying collector number}

\item{coll_date1}{Character string specifying collection data from, e.g., "19 August 1926"}

\item{coll_date2}{Character string specifying collection data from, e.g., "19 August 2018"}

\item{verbose}{Logical}

\item{port}{Default is 4445L}

\item{browserName}{Character string specifying the browser to use, recommended: "chrome"}

\item{collection}{Either "all" or a vector or integers with number corresponding to collection. For a list of collections use function \code{getCollections()}}

\item{syns}{Logical, if TRUE synonyms from MycoBank and IndexFungorum are searched}
}
\value{
\preformatted{      #Returns a dataframe based of the resulting queried records.
}
}
\description{
Enables programmatic interaction with the MyCoPortal web interface to retrieve complete
datasets of sporocarp records from fungaria collections and citizen-science observations.
This function was modified from the \code{mycoportal} function in the rMyCoPortal package (Krah et al. 2019).
See Simpson and Schilling (2020).
}
\examples{
#Query for all Pleurotus records in Minnesota, USA.
mp_query <- mycoportal_tab("Pleurotus", country="United States", state="Minnesota")

}
\references{
\enumerate{
\item Krah FS, Bates S, Miller A. 2019. rMyCoPortal - an R package to interface
with the Mycology Collections Portal. Biodiversity Data Journal 7:e31511.
\item Simpson, H.J., Schilling, J.S. 2020. Using aggregated field collections data
and the novel R package fungarium to investigate fungal fire association. \emph{Mycologia}.
}
}
