% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_update.R
\name{taxon_update}
\alias{taxon_update}
\title{Update scientific names of fungi}
\usage{
taxon_update(
  data,
  taxon_name = "scientificName",
  authorship = "scientificNameAuthorship",
  status_feed = TRUE,
  species_only = TRUE
)
}
\arguments{
\item{data}{Dataframe containing a column of canonical names (e.g. "Pleurotus", "Pleurotus ostreatus") and a column of corresponding authorships (e.g. "(Fr.) P.Kumm.", "(Jacq.) P.Kumm."). Taxa listed in the dataframe can be from any taxonomic rank from kingdom to species; however, there are caveats when updated names for ranks other than species. See Simpson & Schilling (2020).}

\item{taxon_name}{Character string specifying the name of the column containing canonical names. Default is "scientificName" (name used in MyCoPortal datasets).}

\item{authorship}{Character string specifying the name of the column containing authorships. Default is "scientificNameAuthorship" (name used in MyCoPortal datasets).}

\item{status_feed}{Logical specifiying whether the progress of taxa names being queried in the GBIF database should be printed on the console.}

\item{species_only}{Logical specifiying whether records not identified to the species-level should be removed from the dataset.}
}
\value{
The input dataframe with the following output fields appended:
\item{new_name}{currently accepted scientific name (may be the same as the name originally listed in the input file, meaning that the orginal name is currently accepted)}
\item{new_author}{authorship associated with the currently accepted scientific name}
\item{new_full_name}{name and authorship combined into one character string}
\item{new_kingdom}{kingdom classification based on the new scientific name}
\item{new_phylum}{phylum classification based on the new scientific name}
\item{new_class}{class classification based on the new scientific name}
\item{new_order}{order classification based on the new scientific name}
\item{new_family}{family classification based on the new scientific name}
\item{new_genus}{genus associated with the new scientific name}
\item{new_specific_epithet}{specific epithet asscociated with the new scientific name}
\item{confidence}{confidence score (0-100) for match quality of the full scientific name. Ex: "Trametes versicolor (L.) Lloyd" in sporocarp dataset matched with "Trametes versicolor (L.) Lloyd" in GBIF database would give a confidence score of 100.}
\item{match_type}{refers to match type of the canonical name. EXACT means a perfect match. Ex: "Trametes versicolor" in sporocarp dataset matched to "Trametes versicolor" in GBIF database. FUZZY means an imperfect match, likely due to spelling errors. Ex: "Trametes versacolor" in sporocarp dataset matched with "Trametes versicolor" in GBIF database.}
\item{error}{error code for why a name could not be validated or updated. See function code for information on specific error codes.}
}
\description{
Validates or updates the scientific names of fungi, and their associated taxonomic classification, based on currently accepted scientific consensus listed in the \href{https://www.gbif.org/}{GBIF} Backbone Taxonomy database.
}
\details{
Queries the GBIF database for each taxon, via \code\link[taxize]{get_gbifid_}
in the taxize package (Chamberlain and Szocs 2013; Chamberlain et al. 2020), after correcting any erroneous capitalizations of specific epithets (e.g. "Pleurotus Ostreatus").
If a taxon name is mispelled in the input dataset, the best approximate match is selected from the GBIF match results.
See \code{confidence} and \code{match_type} in the "Value" section for more details about approximate matches.
Note that an internet connection is required to retrieve data from the GBIF database.\cr
\cr
If a queried taxon is matched to a GBIF record and that record has "accepted" taxonomic
status, the queried name is "validated" (i.e. the output "new_name" is the same as the queried name).
If the matched GBIF record is a "synonym", the "accepted" record associated with that
synomym is used to "update" the queried taxon (i.e. the ouput "new_name" is different from the queried name).\cr
\cr
If a queried taxon has no GBIF matches or the GBIF match has "doubtful"
taxonomic status, the queried taxon is not validated or updated (i.e. the output "new_name" will be blank) and an error code is output. See \code{error} in "Value" section.\cr
\cr
Function is tailored for large datasets of diverse taxa, like the archived sporocarp collection/observation records accessible through the \href{https://mycoportal.org/}{MyCoPortal}.
However, function will work on any dataset with taxon names and authorship. Becasue processing time can be significant, it may be helpful to breakup extremely large datasets (>50,000 records) and apply this function to each smaller subset.
This ensures that if an error (e.g. loss of internet connection) would occur while the function is running the time lost would be minimized.\cr
\cr
Subspecies in input dataset are automatically updated to the species level. The accepted synonyms of subspecies listed in the GBIF database are the species names.
}
\examples{
MP_data <- mycoportal_tab("Pleurotus")
MP_data_updated <- taxon_update(MP_data)

}
\references{
\enumerate{
\item Scott Chamberlain and Eduard Szocs (2013). taxize - taxonomic search and retrieval in R. F1000Research, 2:191. URL:http://f1000research.com/articles/2-191/v2.
\item Scott Chamberlain, Eduard Szoecs, Zachary Foster, Zebulun Arendsee, Carl
Boettiger, Karthik Ram, Ignasi Bartomeus, John Baumgartner, James O'Donnell,
Jari Oksanen, Bastian Greshake Tzovaras, Philippe Marchand, Vinh Tran, Maëlle
Salmon, Gaopeng Li, and Matthias Grenié. (2020) taxize: Taxonomic information
from around the web. R package version 0.9.95. https://github.com/ropensci/taxize
\item Simpson, H.J., Schilling, J.S. 2020. Using aggregated field collections data
and the novel R package fungarium to investigate fungal fire association. \emph{Mycologia}.
}
}
\author{
Hunter J. Simpson
}
