% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_clado.R
\name{trait_clado}
\alias{trait_clado}
\title{Plot enrichment factors within a cladogram based on taxonomic classification}
\usage{
trait_clado(
  data,
  formula = ~new_kingdom/new_phylum/new_class/new_order/new_family/new_genus/new_name,
  trait_col = "trait_ratio",
  node_color = TRUE,
  show = "all",
  node_all = TRUE,
  ladderize = TRUE,
  continuous = TRUE,
  layout = "circular",
  ...
)
}
\arguments{
\item{data}{Data.frame/data.table containing taxonomic classification variables (e.g. kingdom, phylum, class, etc), and enrichment values.}

\item{formula}{Formula describing how the taxonomic variables in your data.frame should be nested (e.g ~V1/V2/.../Vn). Taxonomic variables must be factors. See \code{ape::as.phylo.formula}.}

\item{trait_col}{Character string specifying the name of the column containing enrichment factors (must be numeric). Default is: "trait_ratio".}

\item{node_color}{Logical. If TRUE (the default), nodes are colored based on the values in trait_col.}

\item{show}{Character string ("all"), or integer, specifying the number of elements to select from the input data when trait_col is ordered. Positive integers select from data that is put in descending order. Negative integers select from data that is put in ascending order. Ex: "100" selects the top 100 elements (e.g., species) with the highest trait_col values; "-100" selects the top 100 elements (e.g., species) with the lowest trait_col values}

\item{node_all}{Logical. If TRUE (the default), all elements in the input data are used for calculating node enrichment values. This becomes important if \code{show} is used to select a subset of the data for display in the cladogram.}

\item{ladderize}{Logical. Should the cladogram be ladderized? Default is TRUE. See \code{ape::ladderize}.}

\item{continuous}{Logical. Should tree coloring be continuous between nodes? Default is TRUE. See \code{ggtree::ggtree}.}

\item{layout}{Character string specifying the type of tree layout. Default is "circular". See \code{ggtree::ggtree}.}

\item{...}{Additional args passed to ggtree. See \code{ggtree::ggtree}.}
}
\value{
\preformatted{      Returns a ggtree object.
}
}
\description{
Taxonomic classification data is used to create a cladogram,
with node colors corresponding to trait enrichment factors. Useful for displaying
the distribution of a trait among different taxonomic groups.
}
\details{
Cladogram produced using \code{ggtree::ggtree};
thus, all \code{ggtree} arguments are accepted. Additional
ggplot layers can be added to the tree as if it were a ggplot object.\cr
\cr
Please cite: \code{ggtree}. To get citation info use: \code{citation(package = 'ggtree')}.\cr
\cr
Cladogram in Simpson and Schilling (2021) created using \code{trait_clado}.
}
\examples{
library(fungarium)
data(strophariaceae) #import sample dataset
data <- taxon_update(strophariaceae, show_status=FALSE) #update taxon names

#Finds fire-associated records
string1 <- "(?i)charred|burn(t|ed)|scorched|fire.?(killed|damaged|scarred)|killed.by.fire"

#Removes records falsely identfied as fire-associated
string2 <- "(?i)un.?burn(t|ed)"

#filter out records that do not contain any environmental metadata (optional)
data <- data[data$occurrenceRemarks!=""|data$host!=""|
                   data$habitat!=""|data$substrate!="",]

#find trait-relevant records
trait_data <- find_trait(data, pos_string=string1, neg_string=string2)

#get trait enrichment
trait_enrichment <- enrichment(all_rec=data, trait_rec=trait_data, status_feed=FALSE)

#filter taxa based on collector bias (optional)
trait_enrichment <- trait_enrichment[trait_enrichment$max_bias<=0.75,]

#filter taxa based on total number of records (optional)
trait_enrichment <- trait_enrichment[trait_enrichment$freq>=5,]

#make cladogram
library(ggtree)
library(ggplot2)

tree <- trait_clado(data=trait_enrichment, trait_col="trait_ratio", continuous=TRUE,
                    ladderize=TRUE, layout="circular", size=1,
                    formula = ~new_order/new_family/new_genus/new_species)+
  geom_tiplab2(color = "black", hjust = 0, offset = 0.1,
               size = 1.5, fontface = "italic") + #add species labels
  geom_tippoint(shape=20, aes(color=trait_ratio, size=trait_freq))+#add tree tips
  ggtitle("Strophariaceae (US records): fire-association")+
  scale_color_gradientn(colours= c("cyan", "blue", "purple", "red", "orange"),
                        name = "Fire-associated records enrichment",
                        limits = c(0, max(trait_enrichment$trait_ratio)+
                                   (0.01*max(trait_enrichment$trait_ratio))),
                        guide = guide_colourbar(label.vjust = 0.6,
                                                label.theme = element_text(size = 10,
                                                                           colour = "black",
                                                                           angle = 0),
                                                title.position = "top",
                                                nbin=100,
                                                draw.ulim = FALSE,
                                                draw.llim = FALSE,
                                                barwidth = 15,
                                                barheight = 0.5)
  )+
  scale_size(name = "Fire-associated records",
             guide = guide_legend(keywidth = 2,
                                  keyheight = 1,
                                  label.position = "bottom",
                                  label.vjust = 0.6,
                                  label.theme = element_text(size = 10,
                                                             colour = "black",
                                                             angle = 0),
                                  title.position = "top")) +
  theme(plot.margin=margin(0,0,0,0),
        legend.title = element_text(size = 10, margin = margin(0,0,0,0)),
        legend.title.align = 0.5,
        legend.position = "bottom",
        legend.justification = "center",
        legend.margin = margin(0,0,0,0),
        plot.title = element_text(hjust = 0.5, margin=margin(0,0,0,0)))+
  xlim(-1, 3.6)#move root away from center; can help improve appearance of circular plot

tree
}
\references{
\enumerate{
\item Simpson, H.J., Schilling, J.S. 2021. Using aggregated field collections data
and the novel R package fungarium to investigate fungal fire association. \emph{Mycologia}. \bold{IN PRESS}
}
}
