% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_funguild.R
\name{assign_funguild}
\alias{assign_funguild}
\title{Assign FUNGuild data}
\usage{
assign_funguild(
  data,
  url = "http://www.stbates.org/funguild_db.php",
  tax_cols = c("phylum_pres", "class_pres", "order_pres", "family_pres", "genus_pres",
    "species_pres")
)
}
\arguments{
\item{data}{\code{dwca} object with updated taxonomy.}

\item{url}{Character. URL where FUNGuild database lives. Current default: http://www.stbates.org/funguild_db.php.}

\item{tax_cols}{Character. Column names containing taxonomic variables. Names must be in descending order of taxonomic rank. Default is c("phylum_pres", "class_pres", "order_pres", "family_pres", "genus_pres", "species_pres").}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Returns the input data.frame with FUNGuild assignments appended.
}\if{html}{\out{</div>}}
}
\description{
Assigns FUNGuild data (e.g. trophic mode, functional guild, etc) to taxa listed in a
dataframe using the FUNGuild database (PLEASE CITE:Nguyen et al. 2016).
}
\examples{
library(fungarium)
data(agaricales_updated) #import sample dataset with updated taxon names
agaricales_updated <- as_dwca(agaricales_updated) # convert to dwca object
agaricales_fg <- assign_funguild(agaricales_updated) #get funguild assignments

}
\references{
\enumerate{
\item Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J, Schilling JS, Kennedy PG. 2016. FUNGuild: An open annotation tool for parsing fungal community datasets by ecological guild. \emph{Fungal Ecology}, 20: 241-248. doi:10.1016/j.funeco.2015.06.006
}
}
