% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grid.R
\name{assign_grid}
\alias{assign_grid}
\title{Overlay grid with equal area cells onto map}
\usage{
assign_grid(
  data,
  lat_col = "lat_parsed",
  lon_col = "lon_parsed",
  size,
  proj = "cea"
)
}
\arguments{
\item{data}{\code{Data.frame}.}

\item{lat_col}{Character. Name of column that contains latitude values.}

\item{lon_col}{Character. Name of column that contains longitude values.}

\item{size}{Integer. Grid cell area (in square kilometers).}

\item{proj}{Character. Equal area map projection to use. Options: 'cea'
(cylindrical equal area) or 'eqearth' (equal earth). Default is 'cea'.}
}
\value{
List containing the following elements:
\describe{
\item{\code{data}}{Input \code{data.frame} with appended \code{grid_cell_id}.}
\item{\code{grid_ref}}{'data.frame' linking 'grid_cell_id' to corresponding grid cell geometries. \code{grid_cell_size} is in sqauare kilometers}
}
}
\description{
Overlays hexagonal grid cells of specified area onto equal area world map.
}
\note{
For more info on the Equal Earth projection see:
https://proj.org/en/stable/operations/projections/eqearth.html.
For cylindrical equal area see: https://proj.org/en/stable/operations/projections/cea.html.
}
\examples{
TODO
}
