% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_grid.R
\name{assign_grid}
\alias{assign_grid}
\title{Overlay grid with equal area cells onto map}
\usage{
assign_grid(
  data,
  lat_col = "lat_parsed",
  lon_col = "lon_parsed",
  size,
  proj = "cea"
)
}
\arguments{
\item{data}{\code{Data.frame}.}

\item{lat_col}{Character. description}

\item{lon_col}{Character. description}

\item{size}{Integer. Grid cell area (in square kilometers).}

\item{proj}{Character. Equal area map projection to use. Options: 'cea'
(cylindrical equal area) or 'eqearth' (equal earth). Default is 'cea'.}
}
\value{
TODO
\describe{
\item{\code{geometry}}{Grid cell geometry.}
\item{\code{grid_cell_id}}{Grid cell ID. Each grid cell has a unique number.}
\item{\code{grid_cell_size}}{The area (in sqauare kilometers) of each grid cell.}
}
}
\description{
Overlays hexagonal grid cells of specified area onto equal area world map.
}
\note{
For more info on the Equal Earth projection see:
https://proj.org/en/stable/operations/projections/eqearth.html.
For cylindrical equal area see: https://proj.org/en/stable/operations/projections/cea.html.
}
\examples{
library(fungarium)
data(agaricales) #import sample data set
clean_geo <- clean_geography(as_dwca(agaricales)) #clean geo
grid_data <- assign_grid(clean_geo, 80000L)
}
