% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abundance.R
\name{calc_abundance}
\alias{calc_abundance}
\title{Calculate abundance}
\usage{
calc_abundance(
  data,
  of = "species_pres",
  per = c("year_parsed", "grid_cell_id"),
  transformation = NULL,
  pseudocount = 0
)
}
\arguments{
\item{data}{Data.frame.}

\item{of}{...}

\item{per}{description}

\item{transformation}{...}

\item{pseudocount}{description}
}
\value{
Input data.frame object with the following fields appended:
\describe{
\item{\code{abund}}{Abundance of each value in the \code{of} field per each unique combination of the \code{per} fields.}
\item{\code{total_occ}}{Total number of occrrences in the field specified by \code{per}.}
\item{\code{rel_abund}}{Relative abundance of each value in the \code{of} field per each unique combination of the \code{per} fields.}
}
}
\description{
Aggregates occurrences based on specified fields to reduce the effects of
spatiotemporal sampling bias. e.g., 50 records for one taxon, at one location,
on the same day, is likely the result of biased sampling. Aggregating these
50 occurrences to one can help reduce this bias.
}
\examples{
library(fungarium)
data(agaricales) #import sample data set
clean_tax <- clean_taxonomy(as_dwca(agaricales)) #clean taxonomy
abund <- calc_abundance(clean_tax)
}
