% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_taxonomy.R
\name{clean_taxonomy}
\alias{clean_taxonomy}
\title{Clean and harmonize taxon names}
\usage{
clean_taxonomy(
  data,
  name_col = "scientificName",
  author_col = "scientificNameAuthorship",
  kingdom = "Fungi",
  refresh_db = FALSE,
  db_url = "https://download.catalogueoflife.org/col/annual/2024_dwca.zip",
  threads = 1L
)
}
\arguments{
\item{data}{\code{Data.frame}.}

\item{name_col}{Character. Column containing taxon names.}

\item{author_col}{Character. Column containing author names.}

\item{kingdom}{Character. \code{Fungi} or \code{Plantae}. Used to speed up taxon matching. Default: \code{Fungi}.}

\item{refresh_db}{Logical. Should the COL database be refreshed. Default: \code{FALSE}. Note, database is always downloaded during the first execution of this function after package installation.}

\item{db_url}{Character. URL for COL database. Default: \verb{https://download.catalogueoflife.org/col/annual/2024_dwca.zip}.}
}
\value{
Input \code{data.frame} with the following output fields appended.

\describe{
\item{\code{kingdom_pres}}{Character. ...}
\item{\code{phylum_pres}}{Character. ...}
\item{\code{class_pres}}{Character. ...}
\item{\code{order_pres}}{Character. ...}
\item{\code{family_pres}}{Character. ...}
\item{\code{genus_pres}}{Character. ...}
\item{\code{specific_epithet_pres}}{Character. ...}
\item{\code{species_pres}}{Character. ...}
\item{\code{species_authority_pres}}{Character. ...}
\item{\code{match_type}}{Character. ...}
\item{\code{match_score}}{Integer. ...}
\item{\code{taxon_raw}}{Character. ...}
\item{\code{taxon_authority_raw}}{Character. ...}
}
}
\description{
Parses taxon names of fungi and plants based on \href{https://www.catalogueoflife.org/}{Catalogue of Life} data.
Synonyms are updated to currently accepted names.
}
\note{
...
}
\examples{
library(fungarium)
data(agaricales) #import sample data set
clean_tax <- clean_taxonomy(agaricales) #clean taxonomy

}
