% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_clean.R
\name{coord_clean}
\alias{coord_clean}
\title{Clean lat/long coordinates in occurrence data sets}
\usage{
coord_clean(
  data,
  lat = "decimalLatitude",
  lon = "decimalLongitude",
  country = "country",
  tests = c("all"),
  centroid_dis = 100,
  round_digits = 4
)
}
\arguments{
\item{data}{Data.frame of occurrence data (e.g. MyCoPortal, GBIF) that includes decimal latitude and longitude}

\item{lat}{Character string specifying the decimal latitude column. Default is "decimalLatitude" (i.e., the Darwin Core standard).}

\item{lon}{Character string specifying the decimal longitude column. Default is "decimalLongitude" (i.e., the Darwin Core standard).}

\item{country}{Character string specifying the name of the country column. Default is "country" (i.e., the Darwin Core standard).}

\item{tests}{Character vector specifying the coordinate cleaning tests to perform. Options include: "zero", "equal", "countries", "centroids". Default is c("zero", "equal", "countries", "centroids").}

\item{centroid_dis}{Numeric specifying the distance threshold (in meters) to use for the centroid test. Default is 100.}

\item{round_digits}{Integer specifying the number of decimal places to use for rounding coordinates. Default is 4. If NULL, no rounding is performed.}
}
\value{
Data.frame containing records from the input data set that passed the coordinate cleaning tests. Number of records removed at each step is printed to the console.
}
\description{
Removes records that do not pass various coordinate validity tests.
}
\details{
The following tests are automatically done:
\describe{
\item{\code{non-numeric}}{lat or lon are not numeric or cannot be converted to numeric}
\item{\code{non-valid}}{lat or lon are invalid numbers (i.e., lat>90, lat<-90, lon>180, lon<-180)}
}
The following tests can be selected:
\describe{
\item{\code{zero}}{lat and lon are both zero}
\item{\code{equal}}{lat and lon are equal}
\item{\code{countries}}{point is outside the bounds of the country listed}
\item{\code{centroid}}{distance between point and country centroid is less than or equal to the centroid_dis specified}
}
}
\examples{
library(fungarium)
data(agaricales) #import sample data set
clean <- coord_clean(agaricales) #clean dates

}
