% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_assign.R
\name{fg_assign}
\alias{fg_assign}
\title{Assign FUNGuild data}
\usage{
fg_assign(
  tax_table,
  url = "http://www.stbates.org/funguild_db.php",
  tax_cols = c("new_phylum", "new_class", "new_order", "new_family", "new_genus",
    "new_name")
)
}
\arguments{
\item{tax_table}{Data.frame/data.table of taxonomic classifications (e.g. phylum, class, order, family, genus, and species names)}

\item{url}{URL where FUNGuild database lives. Current default: http://www.stbates.org/funguild_db.php.}

\item{tax_cols}{Character vector specifying the column names containing taxonomic variables. Names must be in descending order of taxonomic rank.
Default is c("new_phylum", "new_class", "new_order", "new_family", "new_genus", "new_name").}
}
\value{
\preformatted{      Returns the input data.frame/data.table with FUNGuild assignments appended.
}
}
\description{
Assigns FUNGuild data (e.g. trophic mode, functional guild, etc) to taxa listed in a
dataframe using the FUNGuild database (PLEASE CITE:Nguyen et al. 2016).
}
\examples{
library(fungarium)
data(strophariaceae) #import sample dataset
MP_data <- taxon_update(strophariaceae) #update taxon names

#Finds fire-associated records
string1 <- "(?i)charred|burn(t|ed)|scorched|fire.?(killed|damaged|scarred)|killed.by.fire"

#Removes records falsely identfied as fire-associated
string2 <- "(?i)un.?burn(t|ed)"

#filter out records that do not contain any environmental metadata (optional)
MP_data <- MP_data[MP_data$occurrenceRemarks!=""|MP_data$host!=""|
                   MP_data$habitat!=""|MP_data$substrate!="",]

#find trait-relevant records
trait_data <- find_trait(MP_data, pos_string=string1, neg_string=string2)

#get trait enrichment
trait_enrichment <- enrichment(all_rec=MP_data, trait_rec=trait_data)

#filter taxa based on total number of records (optional)
trait_enrichment <- trait_enrichment[trait_enrichment$freq>=5,]

#get funguild assignments
trait_w_funguild <- fg_assign(trait_enrichment)

}
\references{
\enumerate{
\item Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J, Schilling JS, Kennedy PG. 2016. FUNGuild: An open annotation tool for parsing fungal community datasets by ecological guild. \emph{Fungal Ecology}, 20: 241-248. doi:10.1016/j.funeco.2015.06.006
}
}
