% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_assign.R
\name{fg_assign}
\alias{fg_assign}
\title{Assign FUNGuild data}
\usage{
fg_assign(
  tax_table,
  url = "http://www.stbates.org/funguild_db.php",
  phylum = "new_phylum",
  class = "new_class",
  order = "new_order",
  family = "new_family",
  genus = "new_genus",
  species = "new_name"
)
}
\arguments{
\item{tax_table}{Dataframe of taxonomic classifications. Must include columns with phylum, class, order, family, genus, and species names. Missing values is allowed, but may inhibit the assignment of FUNGuild data to that row.}

\item{url}{URL where FUNGuild database lives. Current default: http://www.stbates.org/funguild_db.php.}

\item{phylum}{Character string specifying the phylum column name. Default is "new_phylum" (name given in the output of \code{\link[fungarium]{taxon_update}}).}

\item{class}{Character string specifying the class column name. Default is "new_class".}

\item{order}{Character string specifying the order column name. Default is "new_order".}

\item{family}{Character string specifying the family column name. Default is "new_family".}

\item{genus}{Character string specifying the genus column name. Default is "new_genus".}

\item{species}{Character string specifying the species column name (i.e. full canonical name, not just specific epithet). Default is "new_name".}
}
\value{
\preformatted{      Returns the input dataframe with FUNGuild assignments appended.
}
}
\description{
Assigns FUNGuild data (e.g. trophic mode, functional guild, etc) to taxa listed in a
dataframe using the FUNGuild database (PLEASE CITE:Nguyen et al. 2016).
This is a modifed version of the fg_assign function originally built by
Colin Averill (https://github.com/colinaverill/fg_assign.r).
}
\examples{

mycoportal_data <- taxon_update(mycoportal_data)
guild_data <- fg_assign(mycoportal_data)

}
\references{
\enumerate{
\item Nguyen NH, Song Z, Bates ST, Branco S, Tedersoo L, Menke J, Schilling JS, Kennedy PG. 2016. FUNGuild: An open annotation tool for parsing fungal community datasets by ecological guild. \emph{Fungal Ecology}, 20: 241-248. doi:10.1016/j.funeco.2015.06.006
\item Simpson, H.J., Schilling, J.S. 2020. Using aggregated field collections data and the novel R package “fungarium” to investigate fungal fire association. \emph{Mycologia}.
}
}
