% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_trait.R
\name{find_trait}
\alias{find_trait}
\title{Find fungal records with specific trait data}
\usage{
find_trait(
  data,
  metadata_cols = c("habitat", "occurrenceRemarks", "associatedTaxa"),
  pos_string,
  neg_string = NULL,
  string_clean = TRUE
)
}
\arguments{
\item{data}{\code{Data.frame}.}

\item{metadata_cols}{Character. Names of columns with trait-relevant metadata. Default (c("habitat", "occurrenceRemarks", "associatedTaxa")) is based on fields in \href{https://dwc.tdwg.org/terms/}{Darwin Core archive files} that typically contain trait-relevant metadata.}

\item{pos_string}{Character. String ("positive") containing a regular expression that is used to find strings within the specified metadata columns of the input data frame that contain trait-relevant keywords or phrases.}

\item{neg_string}{Character. String ("negative") containing a regular expression that is used to remove records that were falsely identified, via the "positive" search string, as being trait-relevant. Optional.}

\item{string_clean}{Logical. If TRUE (the default), strings in metadata_cols are "cleaned" prior to trait searching. This includes converting strings to lowercase and removing any punctuation (e.g., periods, commas, question marks, etc.) or extra white space.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Filtered input data that matches search criteria.
}\if{html}{\out{</div>}}
}
\description{
Find trait-relevant records (i.e., records with a particular substrate, host, or habitat association) in a data set of fungal collections/observations
(e.g. \href{https://mycoportal.org}{MyCoPortal} or \href{https://www.gbif.org/}{GBIF} data sets).
}
