% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_trait.R
\name{find_trait}
\alias{find_trait}
\title{Find fungal records with specific trait data}
\usage{
find_trait(
  data,
  pos_string,
  metadata_cols = c("host", "substrate", "habitat", "occurrenceRemarks"),
  neg_string = NULL
)
}
\arguments{
\item{data}{Dataframe containing columns of trait-relevant metadata.}

\item{pos_string}{Character string ("positive") containing a \link{regular expression} that is used to find character strings within the specified metadata columns of the input data frame that contain trait-relevant keywords or phrases.}

\item{metadata_cols}{Character vector containing names of columns with trait-relevant metadata. Default names are "occurrenceRemarks", "host", "habitat", and "substrate" (names used in MyCoPortal datasets).}

\item{neg_string}{Character string ("negative") containing a regular expression that is used to remove records that were falsely identfied, via the "positive" search string, as being trait-relevant. This argument is optional.}
}
\value{
\preformatted{      Returns a dataframe of records with trait-relevant metadata.
}
}
\description{
Find records within a dataset of fungal collections/observations
(e.g. \href{https://mycoportal.org/portal/index.php}{MyCoPortal} dataset)
that have environmental metadata relevant to a particular trait.
Example traits: growing on wood, soil, or leaves;
growing on downed log or standing tree;
growing on or near a specific plant taxon;
growing on or near other fungal taxa;
growing in a habitat affected by a particular disturbance (e.g. wind, fire, logging)
}
\examples{
library(fungarium)
data(strophariaceae) #import sample dataset
MP_data <- taxon_update(strophariaceae) #update taxon names

#Finds fire-associated records
string1 <- "(?i)charred|(?i)burn(t|ed)|(?i)scorched|(?i)fire.?(killed|damaged|scarred)|(?i)killed.by.fire"

#Removes records falsely identfied as fire-associated
string2 <- "(?i)un.?burn(t|ed)"

#filter out records that do not contain any environmental metadata (optional) - may help increase accuracy of enrichment values calculated later
MP_data <- MP_data[MP_data$occurrenceRemarks!=""|MP_data$host!=""|MP_data$habitat!=""|MP_data$substrate!="",]

#find trait-relevant records
trait_data <- find_trait(MP_data, pos_string=string1, neg_string=string2)
}
