% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex_grid.R
\name{hex_grid}
\alias{hex_grid}
\title{Overlay grid on map}
\usage{
hex_grid(size)
}
\arguments{
\item{size}{Integer specifying grid cell area (in square kilometers).}
}
\value{
sf object containing:
\describe{
\item{\code{geometry}}{grid cell geometry}
\item{\code{hex}}{hex number ID. Each grid cell has a unique number}
\item{\code{size}}{the area (in sqauare kilometers) of each grid cell}
}
}
\description{
Overlays hexagonal grid cells of specified area onto equal area world map.
}
\details{
World map projection is cylindrical equal area.
}
\examples{
library(fungarium)

#import world shape file
shp <- rnaturalearth::ne_countries('large', returnclass = "sf")

#transform to cylindrical projection (equal area)
shp <- sf::st_transform(shp, crs = "+proj=cea +ellps=WGS84 +datum=WGS84")

#get grid
grid <- hex_grid(80000)
}
