% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwca.R
\name{import_dwca}
\alias{import_dwca}
\title{Import a DWCA file}
\usage{
import_dwca(file, sep = "\\t", quote = "\\"", select = NULL, drop = NULL)
}
\arguments{
\item{data}{A data.frame with required Darwin Core fields. ("scientificName",
"scientificNameAuthorship", "eventDate", "country", "stateProvince",
"decimalLatitude", "decimalLongitude", "habitat", "occurrenceRemarks",
"associatedTaxa")}
}
\value{
An object of class 'dwca' (and 'data.frame').
}
\description{
Validates and assigns the S3 class 'dwca' to a data.frame.
}
