% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_clado.R
\name{trait_clado}
\alias{trait_clado}
\title{Plot enrichment factors within a cladogram based on taxonomic classification}
\usage{
trait_clado(
  data,
  formula = ~new_kingdom/new_phylum/new_class/new_order/new_family/new_genus/new_name,
  trait_col = "trait_ratio",
  node_color = TRUE,
  filter = NULL,
  node_all = TRUE,
  ladderize = TRUE,
  continuous = TRUE,
  layout = "circular",
  ...
)
}
\arguments{
\item{data}{Data.frame/data.table containing taxonomic classification variables (e.g. kingdom, phylum, class, etc), and enrichment values.}

\item{formula}{Formula describing how the taxonomic variables in your data.frame should be nested (e.g ~V1/V2/.../Vn). Taxonomic variables must be factors. See \code{ape::as.phylo.formula}.}

\item{trait_col}{Character string specifying the name of the column containing enrichment factors (must be numeric). Default is: "trait_ratio".}

\item{node_color}{Logical. If TRUE (the default), nodes are colored based on the values in trait_col.}

\item{filter}{Character vector specifying data set filtering parameters. Default is NULL. To select taxa with the highest or lowest values for a certain variable (e.g., trait_ratio, trait_freq, freq) use "high-100-trait_ratio", "high-150-trait_freq", "low-200-freq", etc. To set a value threshold filter (i.e., filter out taxa with less than or more than a certain value for trait_ratio, trait_freq, etc.) use "trait_freq>=5", "freq>20", etc. Full vector example: c("freq>=10","high-150-trait_ratio"). This example will filter out all taxa with less than 10 trait_freq and then select the top 150 taxa with the highest trait_ratio values. Note that inequality filters are used prior to selecting "high" or "low" taxa.}

\item{node_all}{Logical. If TRUE (the default), all taxa in the original input data (before filtering via the \code{filtering} parameter) are used for calculating node enrichment values. This becomes important if \code{filter} is non NULL.}

\item{ladderize}{Logical. Should the cladogram be ladderized? Default is TRUE. See \code{ape::ladderize}.}

\item{continuous}{Logical. Should tree coloring be continuous between nodes? Default is TRUE. See \code{ggtree::ggtree}.}

\item{layout}{Character string specifying the type of tree layout. Default is "circular". See \code{ggtree::ggtree}.}

\item{...}{Additional args passed to ggtree. See \code{ggtree::ggtree}.}
}
\value{
\preformatted{      Returns a ggtree object.
}
}
\description{
Taxonomic classification data is used to create a cladogram,
with node colors corresponding to trait enrichment factors. Useful for displaying
the distribution of a trait among different taxonomic groups.
}
\details{
Cladogram produced using \code{ggtree::ggtree};
thus, all \code{ggtree} arguments are accepted. Additional
ggplot layers can be added to the tree as if it were a ggplot object.\cr
\cr
Please cite: \code{ggtree}. To get citation info use: \code{citation(package = 'ggtree')}.\cr
\cr
Cladogram in Simpson and Schilling (2021) created using \code{trait_clado}.
}
\examples{
library(fungarium)

#load sample enrichment data set for global Agaricales records
data(agaricales_enrich)

#make circle cladogram
library(ggtree)
library(ggplot2)

#filter out taxa with high collector bias
agaricales_enrich <- agaricales_enrich[agaricales_enrich$max_bias<=0.75,]
agaricales_enrich <- agaricales_enrich[agaricales_enrich$max_bias_t<=0.75,]

#make cladogram
trait_clado(data=agaricales_enrich, trait_col="trait_ratio", continuous=TRUE,
            ladderize=TRUE, layout="circular", size=0.8,
            formula = ~new_order/new_family/new_genus/new_name,
            filter=c("high-300-trait_ratio", "freq>=5")+#make tree
  geom_tiplab2(color = "black", hjust = 0, offset = 0.1,
               size = 1.3, fontface = "italic") + #add species labels
  geom_tippoint(shape=20,
                aes(color=trait_ratio, size=trait_freq),
                alpha=0.75)+#add tree tips
  scale_color_gradientn(colours= c("cyan", "blue", "purple", "red", "orange"),
                        name = "Fire-associated records enrichment",
                        limits = c(0, max(agaricales_enrich$trait_ratio)+
                                     (0.01*max(agaricales_enrich$trait_ratio))),
                        guide = guide_colourbar(label.vjust = 0.6,
                                                label.theme = element_text(size = 10,
                                                                           colour = "black",
                                                                           angle = 0),
                                                title.position = "top",
                                                nbin=100,
                                                draw.ulim = FALSE,
                                                draw.llim = FALSE,
                                                barwidth = 15,
                                                barheight = 0.5))+
  scale_size(name = "Fire-associated records",
             guide = guide_legend(keywidth = 2,
                                  keyheight = 1,
                                  label.position = "bottom",
                                  label.vjust = 0.6,
                                  label.theme = element_text(size = 10,
                                                             colour = "black",
                                                             angle = 0),
                                  title.position = "top")) +
  theme(plot.margin=margin(0,0,0,0),
        legend.title = element_text(size = 10, margin = margin(0,0,0,0)),
        legend.title.align = 0.5,
        legend.position = "bottom",
        legend.justification = "center",
        legend.margin = margin(0,0,0,0),
        plot.title = element_text(hjust = 0.5, margin=margin(0,0,0,0)))+
  xlim(-1,4)#move root away from center;can help improve appearance of circle plot

}
\references{
\enumerate{
\item Simpson, H.J., Schilling, J.S. 2021. Using aggregated field collections data
and the novel R package fungarium to investigate fungal fire association. \emph{Mycologia}. \bold{IN PRESS}
}
}
