% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_clado.R
\name{trait_clado}
\alias{trait_clado}
\title{Plot enrichment factors within a cladogram based on taxonomic classification}
\usage{
trait_clado(
  data,
  formula = ~new_kingdom/new_phylum/new_class/new_order/new_family/new_genus/new_name,
  trait_col = "trait_ratio",
  ladderize = TRUE,
  continuous = TRUE,
  rooted = FALSE,
  layout = "circular",
  ...
)
}
\arguments{
\item{data}{Data.frame/data.table containing taxonomic classification variables (e.g. kingdom, phylum, class, etc), and enrichment values.}

\item{formula}{Formula describing how the taxonomic variables in your data.frame should be nested (e.g ~V1/V2/.../Vn). Taxonomic variables must be factors. See \code{ape::as.phylo.formula}.}

\item{trait_col}{Character string specifying the name of the column containing enrichment factors (must be numeric). Default is: "trait_ratio".}

\item{ladderize}{Logical. Should the cladogram be ladderized? Default is TRUE. See \code{ape::ladderize}.}

\item{continuous}{Logical. Should tree coloring be continuous between nodes? Default is TRUE. See \code{ggtree::ggtree}.}

\item{rooted}{Logical. Should the cladogram be rooted? Default is FALSE.}

\item{layout}{Character string specifying the type of tree layout. Default is "circular". See \code{ggtree::ggtree}.}

\item{...}{Additional args passed to ggtree. See \code{ggtree::ggtree}.}
}
\value{
\preformatted{      Returns a ggtree object.
}
}
\description{
Taxonomic classification data is used to create a cladogram,
with node colors corresponding to trait enrichment factors. Useful for displaying
the distribution of a trait among different taxonomic groups.
}
\details{
Cladogram produced using \code{ggtree::ggtree};
thus, all \code{ggtree} arguments are accepted. Additional
ggplot layers can be added to the tree as if it were a ggplot object.\cr
\cr
Please cite: \code{ggtree}. To get citation info use: \code{citation(package = 'ggtree')}.\cr
\cr
Cladogram in Simpson and Schilling (2021) created using \code{trait_clado}.
}
\examples{
library(fungarium)
data(strophariaceae) #import sample dataset
MP_data <- taxon_update(strophariaceae) #update taxon names

#Finds fire-associated records
string1 <- "(?i)charred|(?i)burn(t|ed)|(?i)scorched|(?i)fire.?(killed|damaged|scarred)|(?i)killed.by.fire"

#Removes records falsely identfied as fire-associated
string2 <- "(?i)un.?burn(t|ed)"

#filter out records that do not contain any environmental metadata (optional) - may help increase accuracy of enrichment values calculated later
MP_data <- MP_data[MP_data$occurrenceRemarks!=""|MP_data$host!=""|MP_data$habitat!=""|MP_data$substrate!="",]

#find trait-relevant records
trait_data <- find_trait(MP_data, pos_string=string1, neg_string=string2)

#get trait enrichment
trait_enrichment <- enrichment(all_rec=MP_data, trait_rec=trait_data)

#filter taxa based on total number of records (optional)
trait_enrichment <- trait_enrichment[trait_enrichment$freq>=5,]

#make cladogram
library(ggtree)
library(ggplot2)

tree <- trait_clado(data=trait_enrichment, trait_col="trait_ratio", continuous=T, ladderize=T, layout="circular", size=0.6)+
 geom_tiplab2(color = "black", hjust = 0, offset = 0.1, size = 1.5, fontface = "italic") +
 ggtitle("Strophariaceae (US records): fire-association")+
 scale_color_gradientn(colours= c("cyan", "blue", "purple", "red", "orange"),
                       name = "Fire-associated records enrichment",
                       limits = c(0, max(trait_enrichment$fire_ratio)),
                       guide = guide_colourbar(label.vjust = 0.6, label.theme = element_text(size = 10, colour = "black", angle = 0),
                                               title.position = "top", nbin=100, draw.ulim = FALSE, draw.llim = FALSE, barwidth = 15, barheight = 0.5)
 )+
 theme(plot.margin=margin(0,0,0,0),
       legend.title = element_text(size = 10, margin = margin(0,0,0,0)),
       legend.title.align = 0.5,
       legend.position = "bottom",
       legend.justification = "center",
       legend.margin = margin(0,0,0,0),
       plot.title = element_text(hjust = 0.5, margin=margin(0,0,0,0)))+
 xlim(-1,6.5)

tree





}
\references{
\enumerate{
\item Simpson, H.J., Schilling, J.S. 2021. Using aggregated field collections data
and the novel R package fungarium to investigate fungal fire association. \emph{Mycologia}. \bold{IN PRESS}
}
}
